Commands now check to prevent empty messages

+++ commands/textreply.js
@@ -13,6 +13,7 @@ module.exports = {
             }
         }
 
+        if (response === "") { return; }
         message.channel.send(response);
 
         message.delete().catch(console.log);
